+++ smartcontract/service/native/shardmgmt/states/states.go
@@ -75,7 +75,7 @@ type ShardState struct {
 	ShardID             types.ShardID  `json:"shard_id"`
 	Creator             common.Address `json:"creator"`
 	State               uint32         `json:"state"`
-	GenesisParentHeight uint64         `json:"genesis_parent_height"`
+	GenesisParentHeight uint32         `json:"genesis_parent_height"`
 	Config              *ShardConfig   `json:"config"`
 	// TODO distinct key lower and upper
 	Peers map[string]*PeerShardStakeInfo `json:"peers"`
